Move clickBehaviorIsValid to metabase-types/guards
It's a predicate over the ClickBehavior wire type (and already delegated
to isImplicitActionClickBehavior, which lives in the guards), not value
formatting - it was only in value-formatting because formatValue happens
to call it. Moving it next to its sibling guard puts it at the lib tier
where the dashboard/parameters consumers and the value-formatting leaf
can all import it, and drops the last conceptually-foreign export from
value-formatting's interface.
