+++ frontend/src/metabase/dashboard/actions/save.ts
@@ -7,12 +7,12 @@ import { createThunkAction } from "metabase/redux";
 import { UPDATE_DASHBOARD_AND_CARDS } from "metabase/redux/dashboard";
 import type { StoreDashboard, StoreDashcard } from "metabase/redux/store";
 import type { Location } from "metabase/router";
-import { clickBehaviorIsValid } from "metabase/value-formatting";
 import type {
   DashCardId,
   ParameterId,
   UpdateCardRequest,
 } from "metabase-types/api";
+import { clickBehaviorIsValid } from "metabase-types/guards";
 
 import { trackDashboardSaved } from "../analytics";
 import { getDashboardBeforeEditing } from "../selectors";
