+++ frontend/src/metabase/value-formatting/index.ts
@@ -1,6 +1,6 @@
 // The module's public interface. Internal helpers stay unexported here on
 // purpose - add a name only when a consumer outside the module needs it.
-export { clickBehaviorIsValid, getDataFromClicked } from "./click-data";
+export { getDataFromClicked } from "./click-data";
 export { displayNameForColumn } from "./column";
 export {
   formatDateTimeForParameter,
